VIC-264 Command line configure wifi for remote user
This should only be for developers and testers and not be exposed to end
users.  This change allows the user on a host to talk to Victor over BLE
and execute a limited number of commands : ifconfig, wpa_cli, dhcptool,
and reboot.  The results of these commands are captured and send back to
the host.

With these tools it is possible to configure WiFi on the Victor over
BLE.  For example:
> wpa_cli add_network
> wpa_cli set_network 1 ssid \"ssidname\"
> wpa_cli set_network 1 psk \"pskvalue\"
> wpa_cli enable_network 1
> wpa_cli save_config

You could then inspect the state of Victor's networking with:
> ifconfig wlan0
> wpa_cli status

I have developed a complementary Linux host tool to act as a shell for
executing these commands and viewing the responses.