VIC-264 Stop advertising once we are connected
Unfortunately, there is no clean callback to tell us when we have
accepted a connection or lost a connection.  Instead, we rely on the
heartbeat failing to send to tell us we lost a connection and on the CCC
descriptor being written to tell us that we accepted a new connection.