VIC-628 Kill background processes on disconnect
The connected central can give us commands to execute on their behalf.
We fork and execute them on a background thread.  If the connected
central disconnects from us, we want to try to kill the currently
running background command and not execute any more commands that are
queued up.

Previously, we were using `android_fork_execvp_ext` to fork and execute
a background process and capture its output.  It worked reasonably well,
however, it didn't give us a way to kill the forked child process.

For this change, I wrote a new set of methods to fork and execute a
background process and capture its output.  I also added a method to
kill the currently executing background process.

Furthermore the methods that execute background commands were modified
to bail out if there isn't a connected central.