VIC-628 Make child exit on failed command
If we try to run a command in the background and execvp fails (likely
because the binary doesn't exist), we need to exit the child process and
print an error message to stderr.  This will allow our parent process,
that is waiting for the result, to get the error and send it back to the
client.