VIC-792 Implement service discovery binder iface
Had to move all of this to the low energy client from the
gatt client because a connection id is required and they are all
stored in the low energy client.

this does the full round trip to ask the stack to search for
services and then receives the callback when the search is
complete.

now i need to figure out how to actually get the list of
services, characteristics, etc.

Change-Id: I8e7e0e4a183f1c1225b6c23cbaee1ed270e273c0