VIC-792 C++ binder plumbing for get_gatt_db
This is the command that will give us the database of services,
characteristics, descriptors, etc. This plumbing is fairly crude in
that it transports a low level structure, btgatt_db_element_t, all the
way from the server side to the client side without too much intelligence
or optimization.  However, the data makes it there and it is usable.