VIC-792 C++ binder glue for char notifications
Allows the client to register for characteristic notifications by mac
address and handle.  Unfortunately, due to a design flaw or low level
bug, the OnCharacteristicNotificationRegistration will NEVER be called.
This is because the client registers by mac address and NOT connection id.
However, the callback is based on connection id.  I had to partially fix
the low level bug by making sure that the callback is passed 0 for the
connection id instead of random stack memory.  Otherwise, we may randomly
look up a connection and call its callback.

Change-Id: I586953e6d0c67783f87ebc7adcda0bac73a6dff4