VIC-862 anki-ble-server is now ankibluetoothd
anki-ble-server used Binder to communicate with Fluoride's
bluetoothtbd daemon to execute BLE commands.  Unfortunately,
bluetoothtbd is not available on our Embedded Linux build and
cannot be easily built and deployed.

ankibluetoothd takes the place of bluetoothtbd and directly
talks to the Bluetooth HAL module (provided by Fluoride).  Various
callbacks are implement to allow us to act as a BLE peripheral the
way we did in anki-ble-server.

In addition, a new script, start_ankibluetoothd, is added that
starts and stops the ankibluetoothd daemon.

Embedded Linux also required build files, configure.ac and
Makefile.am for autotools compatibility.