Do not use QCA wpa_supplicant if 'wifi' is defined
If we are running with 'wifi' feature defined (i. e. our board has
Wi-Fi and we want to play by the ruels), we need to skip inlcuding
QCA's own wpa_supplicant (2.5-git plus extensions) for it not to
stand in the way of stable Yocto's wpa_supplicant (2.4) that we
intend to use.

Change-Id: I06bd9d4b471ab80d3a26a476c917b179f6c40c25