VIC-1210 Crack vicdevsetup off of ankibluetoothd
For security and to enable other developers to access Bluetooth
Low Energy (BLE), ankibluetoothd needs to cleave off the developer
features that support the victor-ble-cli.

Previously, ankibluetoothd directly responded to messages that
came in over 1 of its characteristics and interacted with the
victor-ble-cli to do things like scan for WiFi access points,
configure WiFi, upload SSH keys, etc.  All of those features
are outside the scope of a BLE daemon and represent a security
and stability risk.  Furthermore, these developer features are
NOT how the end user will setup their Victor robot over BLE.

In addition, multiple distinct apps like anki-switchboardd and
anki-cube-controller (names subject to change), will need to
communicate over the BLE interface without having to be concerned
about the low level details of how that interface works.

For these other apps, it is highly desirable to have a simple IPC
mechanism that presents the essential functions of BLE like accepting
a connection, sending/receiving a packet, and disconnecting.

This change takes all of the dangerous and NOT shipable developer
features, that support victor-ble-cli, and puts them in a new app,
vicdevsetup, that talks over IPC to ankibluetoothd to handle
the BLE interface.

--
This patch migrated from:
https://github.com/anki/vic-os--external-anki-ble/tree/129e97adbbad80fddcec81baac27649a957dc90b