Camera mem frame alignment & cleanup
- Ensure camera frame data s 64-byte aligned

The alignment can affect the performance of NEON intrinsics.
Al's original debayer code ensured that memory used for debayer/
downsampling was 64-byte aligned. This change resizes header structs
to be multiples of 64 bytes and ensures that we reserve enough space for
64-byte aligned frame data.

- Allow options to be passed to mm-anki-camera

This defines a path to an environment file that the victor
repo can provide to pass command line options to the camera
server. This will mainly be useful for debugging purposes.

- Display correct camera buf slot in debug log