VIC-1210 anki-ble - Be safer in reading from IPC socket
We are using a stream socket for IPC, so we cannot assume that
a full IPCMessage will be available when we go to read from it. Instead,
we will buffer what we read and when a full IPCMessage is available it
will be processed.

In addition, for safety, we define a maximum message size of 1024 bytes
to prevent attacks and mistakes from blowing up the IPCServer.

Furthermore, sending a message over IPC can now return an error.  In the
future we can use this error reporting to handle dead clients and dead
servers more intelligently.

This commit also fixes a bug with "accepted" sockets being blocking by
using accept4 to make them non-blocking.