VIC-1326 ankibluetoothd: handle dead IPC clients
Our IPC client, currently vicdevsetup, may crash or
die for some reason and then come back to life and reconnect.  In
this change, when an IPC client connects we send it the current
peripheral state so that it can decide what it wants to do.  This allows
BLE connections to be preserved across IPC client deaths.

If we don't want this behavior, we can change `ankibluetoothd` to
drop the connection and clean up state when the IPC client dies.