VIC-1210 ankibluetoothd - support indications
When our BLE peripheral wants to update the connected central of a
new value on a characteristic we need to send either a notification or
an indication.  Notifications are considered unreliable in that they do
not require an acknowledgement from the central.  Indications do.  As
a consequence indications are safer but slower.

This change implements support for the "reliable" parameter of the
SendMessage IPC call by using indications instead of notifications.

vicdevsetup is also modified to send messages reliably.  It is
noticeably slower, but not intolerable.