VIC-1341 Fluoride - Disconnect BLE centrals
When `ankibluetoothd` requests to disconnect a BLE central, a bug
in Fluoride results in the connection staying alive.  I traced this
down the callstack and found that for these inbound connections there
was not a call to change the idle timeout and force a disconnection. I have
added that call along with a comment to explain why it was necessary.