VIC-1476 Victor BLE: Connect to cubes
Before informing our IPC client (like viccubetool) that we have
successfully connected to a BLE peripheral (in this case Victor's
cube), we want to complete service discovery and registering for
notifications on characteristics that support it.

In addition, it was discovered that we have to be really careful when
we ask Fluoride to write descriptors (or presumably characteristics)
and make sure that we don't overflow their internal buffers.  We do
this by keeping our own write queue and only transmit an item after we
get a callback from Fluoride that it is done writing a descriptor or
characteristic.

Furthermore, make sure we don't call the inbound connection callback
for outbound connections.

An additional Fluoride bug occurs when registering for notifications
about a characteristic.  The callback says that the conn_id is 0.  To
workaround this we store the conn_id when we call
register_for_notification.