Customize connman to work with /data (#19)
This patch modifies connmand configuration to expect settings be
located in /data/lib/connman instead of /var/lib/connman. This is
necessary because the root partition will eventually be read-only.

To be date wipeout proof, the default configuration is still
installed into /var/lib/connman and copied to /data/lib/connman
if the latter isn't present. That happens on boot before connman
is launched.

anki-ble writes wifi config to the new location.