VIC-736 VIC-1477 VIC-1479 Flash cube firmware over BLE
To flash cube firmware over BLE, we need to be able to write
characteristics and receive notifications for when readable
characteristics have a new value.  This change introduces the IPC
plumbing for doing that.

In addition, it was observed that it was possible to trigger
congestion warnings when writing characteristics.  Previously, we had
added a safe guard to only attempt to write a characteristic (or
descriptor) after receiving a callback for the previous write attempt.
However, that is insufficient to prevent congestion.  For now, I am
relying on the callback from Fluoride about congestion to stop writing
until the congestion clears.  Hopefully, I do not need to add timer
based throttling.

Flashing firmware is done by writing characteristics and requiring a
response from the cube.  This is slower than writing without a
response, but is much safer.  In testing, writing without response
often produced an error.