VIC-736 Victor BLE : Handle false disconnect notice
Due to a bug in Fluoride, when we create a new connection it ends up
with 2 connection ids.  One for the client interface and one for the
service interface.  When the connection is closed, both callbacks in
the client and server interface get called.  To workaround this, we
need to make sure that the disconnect notice we receive is real and
not for the duplicate connection id.