VIC-1476 Victor BLE : Check connection establishment
When acting as a central and establishing outbound connections,
we want to do the following for our IPC clients before declaring
that the connection is complete:
1. Service Discovery
2. Register with Fluoride for notifications when a notifiable
   characteristic changes
3. Write the CCC descriptors for said characteristics

Steps 2 and 3 depend on step 1, but do not depend on each other.  They can
be done roughly in parallel.  As a result, we need a method to run during
the callbacks for those commands that will check if our connection is
fully established and then we can notify the IPC clients.