wlan/prima: do not register P2P device
The QCA WLAN driver we've got for LE obviously dates back to good
old JellyBean times when Android expected a separate device for
the P2P interface no matter what. So the driver creates that P2P
device sharing the phy with the device in station mode, and even
makes things up in such way that the system thinks it's a station
device too.

So when the time comes to switch to the AP mode,
__wlan_hdd_cfg80211_change_iface called for the funky P2P device
which, again, looks like a station device for the system, and
then the P2P device is removed as this function proceeds, so we
end up changing properties of a deleted device, so we crash in
the context of wpa_supplicant.

As there is no easy way to redo a big part of the driver
architecture and we don't use P2P anyway, let's just comment out
the P2P device creation.