VIC-1574 Set Bluetooth Adapter name from property
The Bluetooth Adapter name is advertised when we broadcast Bluetooth
Low Energy (BLE) advertisements during the pairing process and
normal operation.

The property system from AOSP (part of cutils) provides a simple way
for one process to set values that can be read by other processes in
a reliable way.

With this change, we will use the value in the
`bluetooth.adapter.name` property as the adapter name.  If no value is
provided, use our default value of `VICTOR_<serialno>`.

It is my expectation that some other part of the Victor OS will set
the `bluetooth.adapter.name` property before `ankibluetoothd` is
started and furthermore that this property will NOT change before
`ankibluetoothd` is restarted.