VIC-853 Adds auto white balance correction (#21)
Adds the ability to set the AWB gains. There already existed a sensor process enum SENSOR_SET_AWB_UPDATE however I believe the normal path of execution for it was to be sent by the ISP using mct_events. Since our ISP does not yet work, I have hijacked that existing code path but made it so that it can be executed via the CAM_INTF_PARM enum messages. I have added CAM_INTF_PARM_RAW_AWB_GAIN which is sent via a batchUpdate and eventually triggers a SENSOR_SET_AWB_UPDATE process call. This ends up calling sensor_set_awb_video_hdr_update with an awb_update_t struct as the data. From that the red and blue gains are passed to the camera driver via sensor_fill_awb_array and the correct registers are set to be updated.