VIC-1593 VIC-1606 Christen the robot with a name
Each robot needs a semi-unique name so that it can be distinguished
from others that are nearby.  With this change, on first boot, the
robot will be named Vector XYXY where X is a random letter and Y is a
random digit.  For example : Vector G7H1.  From then on, the name will
not change unless there is a factory reset and the
/data/persist/build.prop file is reset.