VIC-1634 Victor BLE : Fix min/max interval for ads
I discovered that the min/max interval setting is used to add
the Slave Connection Interval Range to the advertisement and
scan response.  This change allows the clients to specify the
min and max interval independently for the advertisement and scan
response.  However, I have also defaulted them to 0 because I don't
believe they are necessary for Victor or its companion app.