VIC-1634 Victor BLE advertising - Cram it all in
We want the end user, whether on an iPhone, Android, or Mac to be able
to quickly scan and find their Victor robot.  When the user presses
the button on Victor, we begin broadcasting BLE advertisements so that
our companion app can find Victor and begin talking to him.  On
iPhone, Android, and Mac these advertisement packets are quickly seen
by their respective BLE stacks and presented to our companion app.  We
need Victor to advertise a few things to the companion app:

1. Connectability - This lets BLE centrals know that they can connect
to Victor.  This takes up 3 bytes in the advertisement packet and is
mandatory.

2. Service UUID - This lets our companion app know it is a Victor and
not some other BLE peipheral.  For a 128-bit UUID, this takes up 18
bytes in the advertisement packet. For a 16-bit UUID, it would only
take up 4 bytes.

3. Name - This name is shown on Victor's face and helps the end user
distinguish their Victor from other nearby Victors.  The user facing
name is 'Vector XYXY'.  It takes up 13 bytes in the advertisement
packet.

4. Manufacturer Specific Data - State information about Victor like
whether he is in pairing mode.  This takes up a variable amount of
space.

The maximum size of an advertising packet is 31 bytes.  Unfortunately,
if you add up the items above you go past 31 bytes.  BLE has a
solution to this, the scan response packet.  When a BLE central is
scanning and seeing our BLE advertisements, it can choose to send a
targeted scan request to an individual BLE peipheral to get an
additional 31 bytes of advertising information.  Unfortunately, the
BLE stacks on some devices are slow to do this or sometimes forget to.
They will pass up the first 31 bytes, but delay or sometimes not even
send up the 31 bytes from the scan response.

Switching to a 16-bit UUID allows us to cram everything into the
31 byte advertisement packet and not worry about the scan response.

To achieve this, I made use of one of Anki's reserved
16-bit Service UUIDs, 0xFEE3 (aka the Anki Single-Message
Service). See the Bluetooth SIG's website for a list of all assigned
16-bit UUIDS:

https://www.bluetooth.com/specifications/assigned-numbers/16-bit-uuids-for-members

We will want to use this 16-bit Service UUID in future Anki products.
To help BLE centrals distinguish which product is advertising,
advertisements will include "Manufacturer Specific Data" with a
product identifier and state information.  In VIC-1639, we are getting
a formally assigned company identifier for manufacturer specific data.
For this commit, I will just use 0xFEE3.