VIC-1680 viccubetool : Flash DVT3 cubes
This change adds the new protocol for flashing DVT3 cubes.
Unfortunately, some very early prototypes incorrectly report that they
are DVT1 cubes.  We will ignore the reported model number and use the
DVT3 flashing protocol.  If an end user needs to flash a DVT1/2 cube
they can use the "flashdvt1" command to do it.  This change also
checks that the cube reports the new firmware version after we flash
it.  Furthermore, we print out the hex bytes that are received on the
app data characteristic rather than converting them to a string which
may be unprintable.