VIC-1752 vicdevsetup : Support WEP, open, and hidden networks
WEP and open networks are terrible insecure things, but unfortunately
they exist in the real world.  Hidden networks, I don't know if they
are more or less secure, but they also exist.  This change allows for
the configuration of all of these types of networks.  The binary
format for describing WiFi networks has changed and a new command code
was added to distinguish from the old wifi-set-config.  As an
additional safeguard, the vic-ble-cli client will have to encode the
name of the SSID in hex.