VIC-1777 Move Bluetooth config files out of /data (#56)
* VIC-1777 Remove bt-app from system image
We don't need Qualcomm's bt-app. It is just for demo purposes.
* VIC-1777 libbt-vendor : Move init.msm.bt.sh to /misc
This static file is generated at build time and should NOT be under
/data as /data can be wiped out.
* VIC-1777 Move Fluoride static files to /misc/bluetooth
The files under /data can be wiped out during an OS update.  Static
configuration files need to be in a persistent place.
* VIC-1777 start_btproperty: mkdir -p /data/misc/bluetooth
The Victor OS image no longer includes files in /data/misc/bluetooth.
Therefore, this directory needs to be created before any process that
depends on it tries to create files there.  The bt-property service
starts up before ankibluetoothd or anything else that uses
/data/misc/bluetooth.