ankibluetoothd - Only write gatt characteristic for outbound
Ooops, the previous logic led to the possibility of trying to writing
a gatt characteristic for the connection id of an inbound connection
(i.e. Victor is the BLE peripheral).  We only want to do that for
connection ids were Victor is the BLE central.  This caused some
extra log spam when sending intense amounts of data (like the contents
of a 6 megabyte file) from Victor to a BLE central.

(cherry picked from commit df27fc933c8cb8ee327416a69a5ceb057dedd76f)