VIC-1820 ankibluetoothd: close IPC socket
If `vicdevsetup` or `switchboard` dies, the IPC socket will be closed.
However, as a result, in ankibluetoothd, we will start getting a storm
of ev::READ events for the closed socket even though we destructed the
watcher.  This leads to a huge CPU spike for `ankibluetoothd` sending
us up to like 25% CPU usage.  To fix this, when we learn that the IPC
socket has been closed by the client, we need to close our end of it
on the server.