VIC-1969 Set hostname and product name to Vector
Our product's name is Vector, not apq8009-robot.  This change updates
the default hostname and the `ro.product.name` property to be Vector.
Furthermore, after ankiinit runs, the hostname will be updated to
the robot's full name, like Vector-XYXY (note the use of '-' instead
of ' ' in the name).

This change will affect the name that the WiFi access point sees and
the DHCP server.  In addition, this will change the product name that
shows up for `adb devices -l`.