VIC-2228 Victor BLE - Connection Parameter Update
We want to change the connection interval between Victor and his
cubes.  This change adds support for calling the connection parameter
update API in the Fluoride stack and shows a simple example of using
it in viccubetool.

In my modification to viccubetool, I set the min and max connection
interval to 20 milliseconds.  Unfortunately, when observing the
interarrival time of notifications from the cube to Victor, I never
saw one that was as short as 20 milliseconds.  I am not sure why that
is.  More experimentation is warranted.