VIC-2037 Remove systemd rescue + emergency targets (#95)
Systemd emergency & rescue targets are intended to handle cases where one of the
normal boot targets has failed by presenting a prompt to the user. We don't support
any kind of interactive prompt, so this doesn't help us at all.  Rather, because there
is no way to access the robot, this mode appears as a "hung" robot. This change removes
both the rescue and emergency target references from target and service files in our
systemd distribution.

It also removes dump and pass (fsck) options from fstab to avoid repairing our partitions.
The reasoning for this is that running fsck on the rootfs will potentially invalidate the image
once dm-verity is enabled.  Running fsck checks on the factory partition at boot is more likely
to cause more problems than it solves.

Our goal is to try as hard as possible to boot.  With these changes, the system continues to
boot even when the factory partition is corrupted.

Brief summary of changes:

- Remove refs to systemd emergency and rescue modes
- Add patch file for removing emergency/rescue targets
- Remove rescue references from ffbm target
- Change fstab to remove dump + fsck (pass) opts