VIC-2071 Remove SELinux from OS images
- VIC-2071 Remove SELINUX and AUDIT support from kernel

- VIC-2071 Remove SELinux from apq8009-robot config

This change removes the userspace components of SELinux from the apq8009-robot machine target.

- VIC-2071 Add back python (removed with selinux)

In our yocto config, selinux appears to come with its own python recipe, which is coupled
tightly to installing selinux. This change adds back python, since we need it for `update-engine`
at the moment.

- VIC-2071 Cleanup SELinux conditionals + prep for ro-rootfs

The configs for apq8053 contain better logic for handling SELinux, and also include
useful changes for turning off journaling on the rootfs, which we might eventually want
to do for shipping images (rootfs will be read-only and devs won't be deploying custom
changes).