msm_serial_hs: work around timer initialization race
The timer structure may not be initialized at the time of call to
mod_timer() since the call is made from the worked thread that is
spawned on probe, while the timer is initialized in
msm_hs_startup(). Work that around by adding per-timer flag
signaling if the timer has been initialized or not.