VIC-2443 Restart BLE ads every 1 minute
For user setup of the Victor device, we need our companion app,
Chewie, to be able to quickly find Victor over BLE and start talking
to him.  Other developers and QA have noticed situations where, after
running for a few minutes or more, Chewie was not seeing the BLE
advertisements from Victor.  I don't know why this is.  I would expect
advertisements to be showing up every few seconds at the worst.  This
is a quick hack to restart the BLE advertisements every minute.  It
worked well enough for me, testing with a Nexus 5 and an iPhone.