VIC-2483 Improve version specification & reporting
It is currently difficult to determine whether an OTA update is starting with the correct
versions. This is because `update-engine` reports only the marketing version (e.g. 0.9.0) and
ignores the incremental version code (build number from TC).

To simplify versioning for the 0.9 release, We now use the `patch` version number to indicate
an incremental build. This makes it easier to identifier versions during OTA updates and keeps
file names simple (e.g 0.9.42). This change also adds an explicit `anki-version.bb` recipe, which
creates files in `/etc` containing version info.  The base version used for all builds now lives
in the `ANKI_VERSION` file in the root of the repo.

- Update ota Makefile to use updated version info and generate a file named:
  `vicos-x.y.z.ota`