VIC-2042 Add tmpfs entries for /run & /var/volatile (#127)
These entries are present in the apq8009/ro-fstab file, which seems to be normally
used for cases where the rootfs is mounted read-only.  In our case, the "robot" variant
is used, so we need to port what we need.

/run is already being mounted as a tmpfs by something (?), so it shouldn't hurt to have
an explicit entry.

In the base-files recipe, /var/volatile is setup to hold:

/tmp -> /var/tmp -> /var/volatile/tmp
/var/log -> /var/volatile/log (used by systemd)

We could probably modify the base-files recipe to change symlinks around to make all
of this work with a single tmpfs, but adding the entries seems to just work correctly
in my testing. I do see systemd files being created in /var/log, so we probably need
to solve this one way or the other. I propose we start with this simple solution and
change it if/when we have evidence that we need to do so.