VIC-2305 Add paycheck.py to Victor OS build
paycheck.py is a script from AOSP under the Apache 2 license:

https://android.googlesource.com/platform/system/update_engine/+/android-8.0.0_r33/scripts/paycheck.py

This script has been slightly modified from the original source to include
an additional option, --do-not-truncate-to-expected-size, that is needed when
applying delta updates on Victor.  On Victor, our partition sizes are currently
larger than the size of the images.  This causes a problem at the end of the update
when the code tries to resize the partition file and fails.  With this new
option, no such attempt is made.