VIC-2304 generate_delta.sh to create image deltas
We want to be able to create small tiny delta files that represent the
changes between an old build and a new one.  This change includes a
binary copy of AOSP's `delta_generator` and the other applications and
shared libraries that it depends on to execute.  `delta_generator` was
built at tag `android-8.0.0_r33`.  To make it easier to use with
Victor OS build artifacts, I have included a wrapper script called
`generate_delta.sh`.