VIC-2607 Victor BLE : IPC API for SetAdapterName
We want to allow vic-switchboard to set the adapter name after it has
set the robot name.  This change adds an IPC API to set the adapter
name at any time.  ankibluetoothd will still try to retrieve the
robot's name to use as the adapter's name on startup.  As before, if
the name cannot be retrieved, a simple name of Vector_<serialno> will
be used.

This change also includes an update to vicdevsetup to show an example
of setting the robot's name if needed (code taken from anki-christen)
and setting the bluetooth adapter's name.