VIC-2657 Add python-pycrypto to Victor OS
We want to use AES encryption for our OTA updates.  Specifically, we
want to use AES in CTR mode.  python-pycrypto has the AES cipher
and the CTR mode available.

ak2539 found a recipe for python-pycrypto at:

http://git.yoctoproject.org/cgit/cgit.cgi/meta-cloud-services/tree/meta-openstack/recipes-devtools/python/python-pycrypto_2.6.1.bb?h=master

Unfortunately, it didn't build under our tree.  Luckily, I was able to
make a simpler recipe that did.  Still had to use the
cross-compiling.patch to get it to build though.

For testing, I deployed the updated system image to the Victor robot
and then launched a python shell.  I was able to use the AES CTR cipher
to encrypt and decrypt a simple message.