Revert "improve roaming" (#153)
Reverts anki/vicos-oelinux#116

We are observing major problems in the office reported by multiple people where robots that associate with an accesspoint (AnkiRobits) drop their connections (dissociate) with re-connecting. Looking through the AP logs, the dissociation / deauthentication appears to be linked to the robots roaming to a new node.

Let's try reverting this PR to prevent the robots from roaming as frequently in an attempt to stop the bleeding.