VIC-2823 Turn on security flags when building vicos
This patch adds a require to the BSP specified security_flags.inc
configuration in local.conf. The security_flags.inc specifies overrides
for CFLAGS and LDFLAGS globally as well as on a per package basis.

Usually the local.conf is not the typical place to add configurations
such as this but it is also not uncommon. The reason this is necessary
is that our BSP is split between two layers that are chosen depending on
priority. This makes it difficult to select the correct
security_flags.inc from the distro specification.