VIC-2753 Configure msm-user distro to not allow GPLv3 packages
This patch adds the INCOMPATIBLE_LICENSE option to the msm-user distro
configuration. When INCOMPATIBLE_LICENSE is set yocto will audit package
licenses to ensure packages with incompatible licenses are not installed.

Yocto will automatically select a package/version that meets the license
requirements. If no such package/version combination exists yocto will fail to
build. If the selected version does not match the preferred version configured
yocto will report the mismatch but will not fail to build.

Currently only the msm-user distro includes the INCOMPATIBLE_LICENSE option in
order to allow development images access to the full range of packages.