VIC-2879 Victor BLE IPC : Suppress SIGPIPE
When we are sending a message to an IPC peer over our Unix Domain
Socket, there is a chance that the peer will exit and close their end
of the socket at just the wrong time.  As a result, a SIGPIPE will be
sent to our process and we will exit.  With this change, each call to
send(2) will have the MSG_NOSIGNAL flag set.  As a result, the SIGPIPE
signal will NOT be generated and instead errno will be set to EPIPE.