VIC-2745 ankibluetoothd: Disconnect unknown peripherals
It is possible that the state ankibluetoothd maintains about
connections to BLE peripherals is out of sync with the state
maintained in Fluoride or the chip itself.  As a result, we could get
notification callbacks from Fluoride for peripherals we
are not tracking.  If we get those callbacks, we want to try to
disconnect immediately.  This will allow later attempts by vic-engine
to have a chance at setting up a proper GATT connection.

(cherry picked from commit baecbb310d51b7bef831774ce21044f841352df2)