VIC-2745 ankibluetoothd: remove_bond hack for disconnect
During my investigation of this issue, on a few occasions, I noticed
that an ACL connection was being formed between the robot and the
cube, but no callbacks for the GATT connection occurred.  Future
attempts to establish GATT connections went down a black hole.  From
reviewing the Fluoride code, I learned that trying to remove a "bond"
should also try to disconnect an ACL connection.  This change adds the
call to remove_bond when we are disconnecting from a peripheral by
address.  This should make it harder for these undesired connections
to exist.  In addition, we also report any unexpected ACL connection
state when trying to establish a GATT connection.

(cherry picked from commit 9505de8ca965a51b783868224799d3dee2ded74a)