VIC-3176 Make chronyd use the Anki NTP pool
chronyd is Victor's NTP client.  We want it to use the Anki NTP pool
to ensure reliable network time.  In addition, it was discovered that
chronyd's configuration was trying to write the drift file and log
files onto the read-only root file system.  This has been corrected to
write files into /data.  As a result, the chronyd service is now
dependent on mount-data.service.