Create FIXTURE flavor (facdev-noverity) for FCC testing
- Add recipe to build menuman from robot/fixture/helpers

For the FIXTURE flavor:

- turn off dm-verity

- set default systemd runlevel to multi-user (not anki-robot)

- Exclude victor.bb recipe from build
  The corresponding victor branch doesn't even build right now...
  Clearly this isn't being used.

- Include anki-menuman.bb recipe instead instead (for menuman)
  This recipe does the following:

  - mkdir /anki/menuman and copy the menuman binary there
  - Copy robot/fixture/helpware/files/* into /anki/menuman

To build this flavor, use the build command:

build-victor-robot-fixture-image

It builds a the equivalent of a "facdev" image with the changes described above.

NOTE: This changeset does not disable ankibluetoothd, since it is currently used
to mark the we booted successfully.  Since dm-verity is disabled, it should be easy
to disable ankibluetoothd by renaming/deleting the binary.