VIC-1678 fake-hwclock - Add in RTC diff at boot
When we reboot Victor, he loads the "current" time from
/data/opt/fake-hwclock/fake-hwclock.data.  But this time is actually
in the past.  It is the time when the VicOS was shutdown.  It is
missing all the seconds between that shutdown and this boot.  As a
result, our clock is not as "current" as we would like.  Luckily, the
RTC is still ticking away when the operating system is shutdown.  By
saving the value of the RTC at shutdown and then seeing the difference
at bootup, we can add enough seconds to our clock to make it closer to
being accurate.  Ultimately, NTP will give us the most accurate time,
but this should help during that period of boot before NTP has
downloaded an accurate time.  One caveat though.  If Victor totally
runs out of battery, the RTC will reset to 0 and we won't add anything
to our saved time.

This change also does a little housekeeping to make sure that we don't
save a partially written file.  It also establishes a baseline epoch
of May 10, 2018 in case the `/etc/timestamp` file is missing.