VIC-3165 Symlink /etc/localtime to /data/etc/localtime
We mount `/` as read only. As a result, `/etc/localtime` cannot be
overwritten to point to a new timezone in /usr/share/zoneinfo.  We
mount `/data` as read/write.  With this change, we symlink
`/etc/localtime` to `/data/etc/localtime` so that the timezone can be
changed.  The default values for `localtime` and `timezone` are
created in `/etc/localtime.default` and `/etc/timezone.default` in
case they are neded.