VIC-3955 Add externalsrc module to victor recipe
The externalsrc module is the prefered method for local sourced packages. It
does a better job checking for changes in the source repository eliminating the
need to always do a clean build. This patch also removes the ota_prod.pub
SRC_URI as it is now included in the victor project at
platform/config/etc/ota.pub