VIC-4071 remove do_shared_work from SRCTREECOVEREDTASKS
Some recipes depend on the kernel's do_shared_workdir task. When the kernel
recipe is set to use externalsrc all of the tasks in SRCTREECOVEREDTASKS are
removed. This patch removes do_shared_workdir from the SRCTREECOVEREDTASKS list
so all recipes depending on the shared_workdir task continue to work when
externalsrc is activated.