VIC-4058 Update linux-quic to use yocto kernel features
This patch coverts the linux-quic kernel recipe to a yocto kernel recipe by
requiring linux-yocto.inc. This was included in order to add support for
externalsrc kernel features and make it easier to upgrade in the future.

This patch is using defconfig files to specify the kernel configuration for
perf and debug builds. The LINUX_KERNEL_TYPE is configured in the distro config
and is used to select the correct defconfig which is now part of the linux-quic
recipe.

In order to update the kernel config copy the
meta/cfg/[perf|debug]/apq8009-robot/defconfig into the kernel as .config and
update it using 'ARCH=arm make menuconfig'. Once modifications are complete
copy the .config back as a defconfig.