VIC-4114 Add qperf to core-image-anki-initramfs recipe
The qperf module handles switching between build types by changing the package
version of every module which inherits qperf. The version is modified to include
the build type. This forces yocto to rebuild artifacts into the directory
structure specific for the current build type. In reality yocto is correct in
that the artifact does not need to be regenerated however we change the deploy
path. In the future it would be ideal to convert the qperf method to a more
yocto friendly approach to build flavors.